+++ libs/estdlib/src/erlang.erl
@@ -52,6 +52,7 @@
     list_to_existing_atom/1,
     list_to_binary/1,
     list_to_integer/1,
+    list_to_integer/2,
     list_to_tuple/1,
     iolist_to_binary/1,
     binary_to_atom/1,
@@ -599,6 +600,19 @@ list_to_binary(_IOList) ->
 %%-----------------------------------------------------------------------------
 -spec list_to_integer(String :: string()) -> integer().
 list_to_integer(_String) ->
+    erlang:nif_error(undefined).
+
+%%-----------------------------------------------------------------------------
+%% @param   String  string to convert to integer
+%% @param   Base  string to convert to integer
+%% @returns an integer value from its string representation
+%% @doc     Convert a string (list of characters) to integer in specified base.
+%% Errors with `badarg' if the string is not a representation of an integer or
+%% the base is out of bounds.
+%% @end
+%%-----------------------------------------------------------------------------
+-spec list_to_integer(String :: string(), Base :: 2..36) -> integer().
+list_to_integer(_String, _Base) ->
     erlang:nif_error(undefined).
 
 %%-----------------------------------------------------------------------------